backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.20.0
joblib==1.3.2
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
Werkzeug==2.3.7
